+++ R/app_server.R
@@ -37,4 +37,5 @@ app_server <- function(input, output, session) {
 
     ## In evidenza
     mod_focus_20200314_server("dapb")
+    mod_focus_20200318_friuli_server("20200318")
 }