+++ R/app_ui.R
@@ -53,7 +53,14 @@ dashboard_sidebar <- function() {dashboardSidebar(sidebarMenu(
 
   menuItem("Home", tabName = "home", icon = icon("home")),
 
-  menuItem("In evidenza", tabName = "focus", icon = icon("bullseye")),
+  menuItem("In evidenza", icon = icon("bullseye"),
+     menuSubItem("2020-03-18 FVG", tabName = "20200318Fvg",
+                 icon = icon("flag")
+     ),
+     menuSubItem("2020-03-14 Veneto", tabName = "20200314Veneto",
+                 icon = icon("flag")
+     )
+  ),
 
   menuItem("Andamento epidemia", icon = icon("chart-line"),
     menuSubItem("Nazionale", tabName = "national", icon = icon("flag")),
@@ -196,9 +203,13 @@ dashboard_body <- function() {dashboardBody(
 
   ),
 
-  tabItem(tabName = "focus",
-    h1("Possibile effetto delle politiche sanitarie in Veneto"),
-    mod_focus_20200314_ui("dapb")
+  tabItem(tabName = "20200318Fvg",
+          h2("Possibile effetto delle politiche sanitarie in Friuli Venezia Giulia"),
+          mod_focus_20200318_friuli_ui("20200318")
+  ),
+  tabItem(tabName = "20200314Veneto",
+          h2("Possibile effetto delle politiche sanitarie in Veneto"),
+          mod_focus_20200314_ui("dapb")
   ),