+++ dev/02_dev.R
@@ -33,6 +33,7 @@ golem::add_module(name = "img_header")
 golem::add_module(name = "help_plot")
 golem::add_module(name = "focus_20200314")
 golem::add_module(name = "ind_ita")
+golem::add_module(name = "focus_20200318_friuli")
 
 
 ## Add helper functions ----