+++ NEWS.md
@@ -1,3 +1,6 @@
+* added module for focus in friuli (20200318)
+* data update (after dpc fix)
+
 # covid19ita 0.12.0
 
 * data update